requirements.txt
ansible
python-digitalocean